Removed redundant PathUtils class
