+++ collect_app/src/main/java/org/odk/collect/android/dao/helpers/ContentResolverHelper.java
@@ -25,13 +25,10 @@ import org.odk.collect.android.logic.FormInfo;
 import org.odk.collect.android.provider.FormsProviderAPI.FormsColumns;
 import org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns;
 import org.odk.collect.android.storage.StoragePathProvider;
-import org.odk.collect.android.storage.StorageSubdirectory;
 
 import java.io.File;
 import java.net.FileNameMap;
 import java.net.URLConnection;
-
-import static org.odk.collect.utilities.PathUtils.getAbsoluteFilePath;
 
 public final class ContentResolverHelper {
 
@@ -73,7 +70,7 @@ public final class ContentResolverHelper {
         try (Cursor c = getContentResolver().query(uri, null, null, null, null)) {
             if (c != null && c.getCount() == 1) {
                 c.moveToFirst();
-                formPath = getAbsoluteFilePath(new StoragePathProvider().getOdkDirPath(StorageSubdirectory.FORMS), c.getString(c.getColumnIndex(FormsColumns.FORM_FILE_PATH)));
+                formPath = new StoragePathProvider().getAbsoluteFormFilePath(c.getString(c.getColumnIndex(FormsColumns.FORM_FILE_PATH)));
             }
         }
         return formPath;
