+++ collect_app/src/main/java/org/odk/collect/android/dao/helpers/FormsDaoHelper.java
@@ -19,9 +19,6 @@ import android.database.Cursor;
 import org.odk.collect.android.dao.FormsDao;
 import org.odk.collect.android.provider.FormsProviderAPI.FormsColumns;
 import org.odk.collect.android.storage.StoragePathProvider;
-import org.odk.collect.android.storage.StorageSubdirectory;
-
-import static org.odk.collect.utilities.PathUtils.getAbsoluteFilePath;
 
 public final class FormsDaoHelper {
 
@@ -45,7 +42,7 @@ public final class FormsDaoHelper {
         try (Cursor c = formsDao.getFormsCursor(selection, selectionArgs)) {
             if (c != null && c.getCount() > 0) {
                 c.moveToFirst();
-                formPath = getAbsoluteFilePath(new StoragePathProvider().getOdkDirPath(StorageSubdirectory.FORMS), c.getString(c.getColumnIndex(FormsColumns.FORM_FILE_PATH)));
+                formPath = new StoragePathProvider().getAbsoluteFormFilePath(c.getString(c.getColumnIndex(FormsColumns.FORM_FILE_PATH)));
             }
         }
         return formPath;
