+++ collect_app/src/main/java/org/odk/collect/android/utilities/FormsDirDiskFormsSynchronizer.java
@@ -24,7 +24,6 @@ import java.util.List;
 import timber.log.Timber;
 
 import static org.odk.collect.android.forms.FormUtils.setupReferenceManagerForForm;
-import static org.odk.collect.utilities.PathUtils.getAbsoluteFilePath;
 
 public class FormsDirDiskFormsSynchronizer implements DiskFormsSynchronizer {
 
@@ -76,7 +75,7 @@ public class FormsDirDiskFormsSynchronizer implements DiskFormsSynchronizer {
                     while (cursor.moveToNext()) {
                         // For each element in the provider, see if the file already exists
                         String sqlFilename =
-                                getAbsoluteFilePath(storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS), cursor.getString(
+                                storagePathProvider.getAbsoluteFormFilePath(cursor.getString(
                                         cursor.getColumnIndex(FormsProviderAPI.FormsColumns.FORM_FILE_PATH)));
                         String md5 = cursor.getString(
                                 cursor.getColumnIndex(FormsProviderAPI.FormsColumns.MD5_HASH));
